Large redesign.
This now uses the latest bootstrap and tabs to separate out the results from the
query information. In addition I changed how the homepage worked. It was
supposed to do something like /jar3d/query-id and create a table with the
sequences. This doesn't work because of the routing so I disabled it. In
addition, we no longer show the refine button on missing queries.

There are also several cosmetic changes, like renaming Templates to templates
and removing the JAR3Doutput folder in templates because it was unneeded.

I've also added a nav bar for jar3d that is inspired by the one for WebR3DAlign.