Update result design
I've moved the jsmol window up to the top of the loops so we can see all
buttons. IN addition, I now use pre blocks for the loop and search boxes as well
as preventing wrapping in these boxes. I've also moved the refine search button
down to near the summary area.