Reorganize the code to be more standard
Django apps have a pretty standard layout so it's a good idea to follow them.
I've moved stuff to match this layout. Mostly this is a bunch of trivial renames
so it isn't a big deal. But we should test and make sure everything still works
before moving it into develop.