Correct query for queue information
This query only got some of the information about the queued work. It didn't get
the group set to use, which is required for doing the comparison to specific
release. This is corrected by getting all of the query information table and the
parts of the corr_queue table that are needed.