First attempt at new beanstalk based queue
This begins the process of adding a beanstalk queue to the webserver.
The current queue is based on perl and is causing issues with running
too long and taking up too much memory. I'm not sure why this is
happening.

I add 2 workers, one for the scoring of an input (bin/score.py) and one
for the aligning of a loop to a model (bin/align.py). These are based
off the generic worker class in JAR3Doutput/worker.py which is a
modification of the worker class from r3d-2-msa. I'm planning on using
subprocess32 as this improves the subprocess in python 2.6 (our current
version). It notably adds timeout support, something we need. The code
is untested right now but this is an outline of what I will attempt to
write.

Once this is done we will remove the queue/ directory entirely.