Add a worker that will score input against motifs
This is the first of 2 workers we need to add that use the queue. This
worker will pull data from the beanstalk queue and then create a java
job to process the data. At the moment this does not correctly update
the database if the jar takes too long. It replaces the old perl based
queue in queue/.